refactor: Remove unused files
